move slider types to options module
