+++ packages/web-components/fast-foundation/src/slider/index.ts
@@ -1,2 +1,3 @@
-export * from "./slider.template.js";
 export * from "./slider.js";
+export * from "./slider.options.js";
+export * from "./slider.template.js";
