+++ packages/web-components/fast-foundation/src/slider/slider.ts
@@ -1,9 +1,4 @@
-import {
-    attr,
-    nullableNumberConverter,
-    observable,
-    SyntheticViewTemplate,
-} from "@microsoft/fast-element";
+import { attr, nullableNumberConverter, observable } from "@microsoft/fast-element";
 import {
     Direction,
     keyArrowDown,
@@ -17,40 +12,7 @@ import {
 import { getDirection } from "../utilities/direction.js";
 import { convertPixelToPercent } from "./slider-utilities.js";
 import { FormAssociatedSlider } from "./slider.form-associated.js";
-
-/**
- * The selection modes of a {@link @microsoft/fast-foundation#(FASTSlider:class)}.
- * @public
- */
-export const SliderMode = {
-    singleValue: "single-value",
-} as const;
-
-/**
- * The types for the selection mode of the slider
- * @public
- */
-export type SliderMode = typeof SliderMode[keyof typeof SliderMode];
-
-/**
- * The configuration structure of {@link @microsoft/fast-foundation#(FASTSlider:class)}.
- * @public
- */
-export interface SliderConfiguration {
-    max: number;
-    min: number;
-    orientation?: Orientation;
-    direction?: Direction;
-    disabled?: boolean;
-}
-
-/**
- * Slider configuration options
- * @public
- */
-export type SliderOptions = {
-    thumb?: string | SyntheticViewTemplate;
-};
+import { SliderConfiguration, SliderMode } from "./slider.options.js";
 
 /**
  * A Slider Custom HTML Element.
